pyproject.toml
[project]
name = "chatbot-pydanticai-template"
version = "0.1.0"
description = "Template for a custom chatbot based on Pydantic AI with integrated UI and extension points."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=1.81.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic-deep>=0.3.13",
    # BPMN Pipeline — LLM provider SDKs
    "anthropic>=0.50.0",
    "google-generativeai>=0.8.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
# Optional: observability (no token = no-op)
logfire = ["logfire>=0.21.0"]
# Optional: MCP tools (Model Context Protocol) - check pydantic-ai docs for exact extra name
# mcp = ["pydantic-ai[mcp]>=0.0.20"]
# Optional: vector store / RAG (add your preferred backend, e.g. chromadb, qdrant)
# vector = ["chromadb>=0.4.0"]

[build-system]
requires = ["uv_build>=0.6,<0.12"]
build-backend = "uv_build"

[tool.uv]
# Install: uv sync
# Run app: uv run python -m chatbot (or uv run uvicorn chatbot.app:app --reload)

[tool.uv.build-backend]
module-name = "chatbot"

[dependency-groups]
dev = ["ty>=0.0.29"]

[tool.ty.environment]
python-version = "3.10"

[tool.ty.rules]
# Enable all error-level rules (default). Adjust as needed:
# possibly-unbound = "warn"
# unresolved-import = "warn"

[tool.ruff]
line-length = 100
target-version = "py310"


[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = [
    "E501",   # line too long
]
